sentry: normalize nullable SDK configuration once
Evaluate nullable organization, sampling, profile, and log-flush environment values once before casting, preserving Sentry's exact null and sample-rate defaults.

Use the Redis capability boundary for topology-specific indexes and rely on the shared environment-test helper so configuration tests exercise scalar normalization without maintaining their own process-state cleanup.

+++ src/sentry/src/Features/RedisFeature.php
@@ -92,7 +92,7 @@ class RedisFeature extends Feature
             'db.system' => 'redis',
             'db.statement' => $redisStatement,
             'db.redis.connection' => $event->connectionName,
-            'db.redis.database_index' => (int) ($config['database'] ?? 0),
+            'db.redis.database_index' => $config['database'] ?? 0,
             'db.redis.pool.name' => $event->connectionName,
             'db.redis.pool.max' => $pool->getOption()->getMaxConnections(),
             'db.redis.pool.max_idle_time' => $pool->getOption()->getMaxIdleTime(),
